reorg: move the IsWrapped stuff alongside GIsomorphic for posterity
